+++ system/interfaces/interfaces.go
@@ -336,9 +336,9 @@ func (ii *Interfaces) SetDoorPasscodes(controller types.IController, door uint8,
 	}
 }
 
-func (ii *Interfaces) PutCard(controller types.IController, card uint32, PIN uint32, from, to lib.Date, permissions map[uint8]uint8) {
+func (ii *Interfaces) PutCard(controller types.IController, card uint32, PIN uint32, from, to lib.Date, permissions map[uint8]uint8, firstcard map[uint8]bool) {
 	if lan, ok := ii.LAN(); ok {
-		lan.putCard(controller, card, PIN, from, to, permissions)
+		lan.putCard(controller, card, PIN, from, to, permissions, firstcard)
 	}
 }